+++ pwnlib/libcdb.py
@@ -11,6 +11,7 @@ import tempfile
 
 from pwnlib.context import context
 from pwnlib.elf import ELF
+from pwnlib.filesystem.path import Path
 from pwnlib.log import getLogger
 from pwnlib.tubes.process import process
 from pwnlib.util.fiddling import enhex
@@ -126,9 +127,28 @@ def provider_local_system(hex_encoded_id, hash_type):
         return local_libc.data
     return None
 
-PROVIDERS = [provider_local_system, provider_libcdb, provider_libc_rip]
+# Offline search https://github.com/niklasb/libc-database for hash type
+def provider_local_database(hex_encoded_id, hash_type):
+    if not context.local_libcdb:
+        return None
 
-def search_by_hash(hex_encoded_id, hash_type='build_id', unstrip=True):
+    localdb = Path(context.local_libcdb)
+    if not localdb.is_dir():
+        return None
+
+    log.debug("Searching local libc database, %s: %s", hash_type, hex_encoded_id)
+    for libc_path in localdb.rglob("*.so"):
+        if hex_encoded_id == HASHES[hash_type](libc_path):
+            return read(libc_path)
+
+    return None
+
+PROVIDERS = {
+    "offline": [provider_local_system, provider_local_database],
+    "online": [provider_libcdb, provider_libc_rip]
+}
+
+def search_by_hash(hex_encoded_id, hash_type='build_id', unstrip=True, offline_only=False):
     assert hash_type in HASHES, hash_type
 
     # Ensure that the libcdb cache directory exists
@@ -140,8 +160,12 @@ def search_by_hash(hex_encoded_id, hash_type='build_id', unstrip=True):
     if cache is None:
         return None
 
+    providers = PROVIDERS["offline"]
+    if not offline_only:
+        providers += PROVIDERS["online"]
+
     # Run through all available libc database providers to see if we have a match.
-    for provider in PROVIDERS:
+    for provider in providers:
         data = provider(hex_encoded_id, hash_type)
         if data and data.startswith(b'\x7FELF'):
             break
@@ -273,6 +297,10 @@ def unstrip_libc(filename):
     if not libc.buildid:
         log.warn_once('Given libc does not have a buildid. Cannot look for debuginfo to unstrip.')
         return False
+
+    if libc.debuginfo:
+        log.debug('Given libc already contains debug information. Skipping unstrip.')
+        return True
 
     log.debug('Trying debuginfod servers: %r', DEBUGINFOD_SERVERS)
 
@@ -586,7 +614,7 @@ def search_by_symbol_offsets(symbols, select_index=None, unstrip=True, return_as
     selected_libc = _handle_multiple_matching_libcs(matching_libcs)
     return search_by_build_id(selected_libc['buildid'], unstrip=unstrip)
 
-def search_by_build_id(hex_encoded_id, unstrip=True):
+def search_by_build_id(hex_encoded_id, unstrip=True, offline_only=False):
     """
     Given a hex-encoded Build ID, attempt to download a matching libc from libcdb.
 
@@ -595,6 +623,10 @@ def search_by_build_id(hex_encoded_id, unstrip=True):
             Hex-encoded Build ID (e.g. 'ABCDEF...') of the library
         unstrip(bool):
             Try to fetch debug info for the libc and apply it to the downloaded file.
+        offline_only(bool):
+            Both offline and online providers are used by default. When pass
+            `offline_only=True`, libcdb enable an exclusive offline search mode,
+            which will disable online providers.
 
     Returns:
         Path to the downloaded library on disk, or :const:`None`.
@@ -610,9 +642,9 @@ def search_by_build_id(hex_encoded_id, unstrip=True):
         >>> hex(ELF(filename).symbols.read)
         '0xeef40'
     """
-    return search_by_hash(hex_encoded_id, 'build_id', unstrip)
+    return search_by_hash(hex_encoded_id, 'build_id', unstrip, offline_only)
 
-def search_by_md5(hex_encoded_id, unstrip=True):
+def search_by_md5(hex_encoded_id, unstrip=True, offline_only=False):
     """
     Given a hex-encoded md5sum, attempt to download a matching libc from libcdb.
 
@@ -621,6 +653,10 @@ def search_by_md5(hex_encoded_id, unstrip=True):
             Hex-encoded md5sum (e.g. 'ABCDEF...') of the library
         unstrip(bool):
             Try to fetch debug info for the libc and apply it to the downloaded file.
+        offline_only(bool):
+            Both offline and online providers are used by default. When pass
+            `offline_only=True`, libcdb enable an exclusive offline search mode,
+            which will disable online providers.
 
     Returns:
         Path to the downloaded library on disk, or :const:`None`.
@@ -636,9 +672,9 @@ def search_by_md5(hex_encoded_id, unstrip=True):
         >>> hex(ELF(filename).symbols.read)
         '0xeef40'
     """
-    return search_by_hash(hex_encoded_id, 'md5', unstrip)
+    return search_by_hash(hex_encoded_id, 'md5', unstrip, offline_only)
 
-def search_by_sha1(hex_encoded_id, unstrip=True):
+def search_by_sha1(hex_encoded_id, unstrip=True, offline_only=False):
     """
     Given a hex-encoded sha1, attempt to download a matching libc from libcdb.
 
@@ -647,6 +683,10 @@ def search_by_sha1(hex_encoded_id, unstrip=True):
             Hex-encoded sha1sum (e.g. 'ABCDEF...') of the library
         unstrip(bool):
             Try to fetch debug info for the libc and apply it to the downloaded file.
+        offline_only(bool):
+            Both offline and online providers are used by default. When pass
+            `offline_only=True`, libcdb enable an exclusive offline search mode,
+            which will disable online providers.
 
     Returns:
         Path to the downloaded library on disk, or :const:`None`.
@@ -662,10 +702,9 @@ def search_by_sha1(hex_encoded_id, unstrip=True):
         >>> hex(ELF(filename).symbols.read)
         '0xeef40'
     """
-    return search_by_hash(hex_encoded_id, 'sha1', unstrip)
+    return search_by_hash(hex_encoded_id, 'sha1', unstrip, offline_only)
 
-
-def search_by_sha256(hex_encoded_id, unstrip=True):
+def search_by_sha256(hex_encoded_id, unstrip=True, offline_only=False):
     """
     Given a hex-encoded sha256, attempt to download a matching libc from libcdb.
 
@@ -674,6 +713,10 @@ def search_by_sha256(hex_encoded_id, unstrip=True):
             Hex-encoded sha256sum (e.g. 'ABCDEF...') of the library
         unstrip(bool):
             Try to fetch debug info for the libc and apply it to the downloaded file.
+        offline_only(bool):
+            Both offline and online providers are used by default. When pass
+            `offline_only=True`, libcdb enable an exclusive offline search mode,
+            which will disable online providers.
 
     Returns:
         Path to the downloaded library on disk, or :const:`None`.
@@ -689,7 +732,7 @@ def search_by_sha256(hex_encoded_id, unstrip=True):
         >>> hex(ELF(filename).symbols.read)
         '0xeef40'
     """
-    return search_by_hash(hex_encoded_id, 'sha256', unstrip)
+    return search_by_hash(hex_encoded_id, 'sha256', unstrip, offline_only)